Moved IdentityModels to Data with corresponding code, installed NuGet packages
