+++ MousePark.Services/packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="EntityFramework" version="6.1.0" targetFramework="net472" />
+  <package id="Microsoft.AspNet.Identity.Core" version="2.2.3" targetFramework="net472" />
+  <package id="Microsoft.AspNet.Identity.EntityFramework" version="2.2.3" targetFramework="net472" />
+</packages>
