+++ MouseParkApi/App_Start/IdentityConfig.cs
@@ -3,6 +3,7 @@ using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
 using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin;
+using MousePark.Data;
 using MouseParkApi.Models;
 
 namespace MouseParkApi
