+++ MouseParkApi/Controllers/AccountController.cs
@@ -13,6 +13,7 @@ using Microsoft.AspNet.Identity.Owin;
 using Microsoft.Owin.Security;
 using Microsoft.Owin.Security.Cookies;
 using Microsoft.Owin.Security.OAuth;
+using MousePark.Data;
 using MouseParkApi.Models;
 using MouseParkApi.Providers;
 using MouseParkApi.Results;
